refactor: split formula plot into modules
